pyproject.toml
[project]
name = "langchain-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.1.0",
    "fastapi>=0.129.2",
    "flask>=3.1.3",
    "isort>=8.0.0",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.10",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]
